requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
